Renamed src/masschange/ingest/utils/flagfield.py -> src/masschange/ingest/utils/arraylikefields.py; made dataproduct.populate_flag() more generic to work with differently named array-like data
